First draft of a XML parser and converter written in Ruby
I've searched for an XML to Markdown parser in Ruby, Javascript,
XSLT and others languages, but I still have not find a library that
handles HTML to Markdown tables properly.

Most Markdown converters fail to understand that tables come with
headers ! They are NEEDED for the comprehension of the table.

Still, I'd rather avoid writing such converters. There should be an
easier way somehow...

Signed-off-by: Myy <myy@miouyouyou.fr>